+++ src/main/kotlin/dev/aaronhowser/mods/irregular_implements/block/block_entity/EnderAnchorBlockEntity.kt
@@ -1,16 +1,68 @@
 package dev.aaronhowser.mods.irregular_implements.block.block_entity
 
-import dev.aaronhowser.mods.irregular_implements.block.MoonPhaseDetectorBlock.Companion.INVERTED
-import dev.aaronhowser.mods.irregular_implements.block.MoonPhaseDetectorBlock.Companion.POWER
+import dev.aaronhowser.mods.irregular_implements.client.render.CubeIndicatorRenderer
 import dev.aaronhowser.mods.irregular_implements.registry.ModBlockEntityTypes
+import dev.aaronhowser.mods.irregular_implements.registry.ModItems
+import dev.aaronhowser.mods.irregular_implements.util.ClientUtil
 import net.minecraft.core.BlockPos
+import net.minecraft.util.Mth
 import net.minecraft.world.level.Level
 import net.minecraft.world.level.block.entity.BlockEntity
 import net.minecraft.world.level.block.state.BlockState
+import kotlin.math.sin
 
 class EnderAnchorBlockEntity(
 	pos: BlockPos,
 	blockState: BlockState
 ) : BlockEntity(ModBlockEntityTypes.ENDER_ANCHOR.get(), pos, blockState) {
+
+	companion object {
+		fun tick(
+			level: Level,
+			pos: BlockPos,
+			state: BlockState,
+			blockEntity: EnderAnchorBlockEntity
+		) {
+			if (!level.isClientSide) return
+
+			val localPlayer = ClientUtil.localPlayer ?: return
+			if (!localPlayer.isHolding(ModItems.PORTABLE_ENDER_BRIDGE.get())) return
+
+			val colorOne = 0xFF14173E
+			val colorTwo = 0xFF383993
+
+			val tick = level.gameTime
+
+			val period = 20.0 * 10.0
+
+			val t = (sin((tick % period) / period * Math.PI * 2.0) + 1.0) / 2.0
+
+			val r = Mth.lerp(
+				t.toFloat(),
+				(colorOne shr 16 and 0xFF).toFloat(),
+				(colorTwo shr 16 and 0xFF).toFloat()
+			).toInt()
+
+			val g = Mth.lerp(
+				t.toFloat(),
+				(colorOne shr 8 and 0xFF).toFloat(),
+				(colorTwo shr 8 and 0xFF).toFloat()
+			).toInt()
+
+			val b = Mth.lerp(
+				t.toFloat(),
+				(colorOne and 0xFF).toFloat(),
+				(colorTwo and 0xFF).toFloat()
+			).toInt()
+
+			val color = (0x77 shl 24) or (r shl 16) or (g shl 8) or b
+
+			CubeIndicatorRenderer.addIndicator(
+				pos,
+				2,
+				color
+			)
+		}
+	}
 
 }