+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 	- Currently no non-footwear items are in that tag, but now you can add them yourself
 - The Biome Painter now defaults to having a radius of 2 instead of 0
   - Now it has a default volume of 5x5x5 instead of 1x1x1
+- When holding a Portable Ender Bridge, Ender Anchors now glow through blocks
 
 ### Fixed